new-backend/talent-intelligence-service/requirements.txt

fastapi
uvicorn
openai
pydantic
streamlit
requests
python-multipart
pdfplumber
python-docx
python-dotenv
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
colorama==0.4.6
cryptography==46.0.5
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.3.1
frozenlist==1.8.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.0
langchain==0.1.16
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.1.7
langchain-text-splitters==0.0.2
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.1.147
lxml==6.0.2
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==1.26.4
openai==1.109.1
orjson==3.11.7
ormsgpack==1.12.2
packaging==23.2
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
propcache==0.4.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.27.2.2
pypdfium2==5.6.0
python-docx==1.2.0
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.48
tenacity==8.5.0
tiktoken==0.12.0
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0